fix(v1): delete item message
