+++ src/buttons/ticket-v1/delete.ts
@@ -1,9 +1,12 @@
 import Ticket from '../../schemas/ticketSchema';
 import { ButtonInteraction } from '../../structures/Command';
 
+/**
+ * @desceiption 배틀이 V1 - 티켓 7월 30일까지만 지원
+  */
 export default new ButtonInteraction(
   {
-    name: 'ticket.delete',
+    name: 'delete',
   },
   async (client, interaction) => {
     await Ticket.updateOne(
